requirements.txt
Django==3.2.6
djangorestframework==3.12.4
django-extensions==3.1.3
psycopg2-binary>=2.8
requests==2.26.0